Introduce `AbstractConsumer`
Cleans up quite a bit of code in consumer and makes the tests a bit
understandable.
